pyproject.toml
[project]
name = "haystack-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "haystack-ai>=2.13.1",
    "markdown-it-py>=3.0.0",
    "mdit-plain>=1.0.1",
    "openai>=1.76.0",
    "qdrant-haystack>=9.1.1",
    "sentence-transformers>=4.1.0",
]
